Ocs_update

+++ src/main/java/org/slavik/OCS/model/Quantity.java
@@ -3,27 +3,20 @@ package org.slavik.OCS.model;
 
 import com.fasterxml.jackson.annotation.*;
 
+import java.time.OffsetDateTime;
+
 public class Quantity {
-    private long value;
+   private int value;
     private boolean isGreatThan;
 
     @JsonProperty("value")
-    public long getValue() {
-        return value;
-    }
-
+    public int getValue() { return value; }
     @JsonProperty("value")
-    public void setValue(long value) {
-        this.value = value;
-    }
+    public void setValue(int value) { this.value = value; }
+
 
     @JsonProperty("isGreatThan")
-    public boolean getIsGreatThan() {
-        return isGreatThan;
-    }
-
+    public boolean getIsGreatThan() { return isGreatThan; }
     @JsonProperty("isGreatThan")
-    public void setIsGreatThan(boolean value) {
-        this.isGreatThan = value;
-    }
+    public void setIsGreatThan(boolean value) { this.isGreatThan = value; }
 }
